optimize dependencies
