+++ CppImport/src/macro/StandardMetaDefinitions.cpp
@@ -26,13 +26,18 @@
 
 #include "StandardMetaDefinitions.h"
 
-#include "MacroExpansion.h"
-#include "MacroArgumentInfo.h"
+#include "ClangHelpers.h"
 #include "MacroDefinitions.h"
+#include "MacroExpansion.h"
 #include "MacroExpansions.h"
+#include "MacroArgumentInfo.h"
+#include "MacroArgumentLocation.h"
 #include "LexicalTransformations.h"
+#include "NodeToCloneMap.h"
+
 #include "NodeHelpers.h"
-#include "AllMetaDefinitions.h"
+
+#include "OOModel/src/allOOModelNodes.h"
 
 namespace CppImport {
